+++ user_interface.h
@@ -13,7 +13,7 @@ enum UserModes{ADMIN=1,PATIENT=0,NONE=-1};
 typedef enum UserModes UserModes;
 
 //function to choose the mode of operation
-u8 modeSet(u8 mode,UserModes* USER_STATUS);
+u32 modeSet(u32 mode,UserModes* USER_STATUS);
 
 //boolean check to see if the admin password was accepted or not
 enum AdminStatus{A_ACCEPTED=1,A_REJECTED=0};
@@ -25,7 +25,7 @@ AdminStatus adminCheck();
 
 //function to compare 2 strings
 //returns 1 for equal strings and 0 for unequal strings
-u8 cmpS(const u8* str1,const u8* str2);
+u32 cmpS(const u8* str1,const u8* str2);
 
 
 //main runtime window for admin user
